+++ asf_mission_data/pipeline/energy_price_cap_levels_annex_9/bronze.py
@@ -1,12 +1,9 @@
 import logging
 import re
-from datetime import datetime
 from pathlib import Path
-from typing import Type
 from urllib.parse import urljoin
 
 from bs4 import BeautifulSoup
-from hamilton.data_quality.base import DataValidator, ValidationResult
 from hamilton.function_modifiers import check_output_custom
 
 from asf_mission_data import storage, utils
@@ -14,6 +11,7 @@ from asf_mission_data.pipeline.energy_price_cap_levels_annex_9.config import (
     PRICE_CAP_PERIOD_PUBLICATION_DATES,
     PRICE_CAP_PERIOD_STRING_PATTERN,
 )
+from asf_mission_data.pipeline.energy_price_cap_levels_annex_9.validators import LatestPriceCapFileUrlValidator, LatestPriceCapValidator
 
 logger = logging.getLogger(__name__)
 
@@ -28,76 +26,6 @@ def latest_collection_page_html_soup(collection_url: str) -> BeautifulSoup:
         BeautifulSoup: Parsed HTML content of the page.
     """
     return BeautifulSoup(utils.fetch_raw_content(collection_url), "html.parser")
-
-
-# TODO move to a more appropriate module later
-class LatestPriceCapFileUrlValidator(DataValidator):
-    """Checks that the price cap file url retrieved is the expected one.
-
-    This is based on the ofgem url structure including the publication month and year.
-    """
-
-    def __init__(
-        self,
-        price_cap_period_publication_dates: dict[str, str],
-        importance: str = "fail",
-    ):
-        super(LatestPriceCapFileUrlValidator, self).__init__(importance=importance)
-        self.price_cap_period_publication_dates = price_cap_period_publication_dates
-
-    def applies_to(self, datatype: Type) -> bool:
-        """Whether or not this data validator can apply to
-        the specified dataset
-
-         :param datatype:
-         :return: True if it can be run on the specified type.
-        """
-        return datatype is str
-
-    def description(self) -> str:
-        """Gives a description of this validator.
-        :return: The description of the validator as a string
-        """
-        return "Checks that the price cap file url retrieved is the expected one given the execution date."
-
-    @classmethod
-    def name(cls) -> str:
-        """Returns the name for this validator."""
-        return "LatestPriceCapFileUrlValidator"
-
-    def validate(self, data: str) -> ValidationResult:
-        """Actually performs the validation.
-
-        :param data: data to validate
-        :return: The result of validation
-        """
-        # Anticipated date pattern is YYYY-MM
-        match = re.search(r"(\d{4}-\d{2})", data)
-
-        if match:
-            extracted_date = match.group(1)
-        else:
-            return ValidationResult(
-                passes=False,
-                message=f"Unexpected price cap file url, no date component in: {data}",
-            )
-
-        # Get expected publication dates is same format as file url date
-        PUBLICATION_DATES = (datetime.fromisoformat(v) for v in self.price_cap_period_publication_dates.values())
-
-        # Get expected price cap period
-        now = datetime.now()
-        latest_publication_date = max((v for v in PUBLICATION_DATES if v <= now))
-
-        # Match anticipated date format
-        latest_publication_date_str = latest_publication_date.strftime(format="%Y-%m")
-
-        valid = extracted_date == latest_publication_date_str
-
-        return ValidationResult(
-            passes=valid,
-            message=f"Expected file url date: {latest_publication_date_str}, saw date {extracted_date}",
-        )
 
 
 @check_output_custom(LatestPriceCapFileUrlValidator(PRICE_CAP_PERIOD_PUBLICATION_DATES))
@@ -128,76 +56,6 @@ def latest_file_url(
             return urljoin(collection_url, a["href"])
 
     raise ValueError(f"Could not find dataset '{file_link_text}' at {collection_url}")
-
-
-# TODO move to a more appropriate module later
-class LatestPriceCapValidator(DataValidator):
-    """Checks that the price cap period retrieved is the expected one."""
-
-    def __init__(
-        self,
-        price_cap_period_publication_dates: dict[str, str],
-        importance: str = "fail",
-    ):
-        super(LatestPriceCapValidator, self).__init__(importance=importance)
-        self.price_cap_period_publication_dates = price_cap_period_publication_dates
-
-    def applies_to(self, datatype: Type) -> bool:
-        """Whether or not this data validator can apply to
-        the specified dataset
-
-         :param datatype:
-         :return: True if it can be run on the specified type.
-        """
-        return datatype is str
-
-    def description(self) -> str:
-        """Gives a description of this validator.
-        :return: The description of the validator as a string
-        """
-        return "Checks that the price cap period retrieved is the expected one given the execution date."
-
-    @classmethod
-    def name(cls) -> str:
-        """Returns the name for this validator."""
-        return "LatestPriceCapValidator"
-
-    def validate(self, data: str) -> ValidationResult:
-        """Actually performs the validation.
-
-        :param data: data to validate
-        :return: The result of validation
-        """
-        extracted_period_interval = utils.convert_energy_price_cap_period_string_to_interval(data)
-        INTERVAL_PUBLICATION_DATES = {
-            utils.convert_energy_price_cap_period_string_to_interval(k): datetime.fromisoformat(v)
-            for k, v in self.price_cap_period_publication_dates.items()
-        }
-
-        if extracted_period_interval not in INTERVAL_PUBLICATION_DATES:
-            return ValidationResult(
-                passes=False,
-                message=f"Unrecognised price cap period: {extracted_period_interval}",
-            )
-
-        # Look up publication date of extracted price cap period
-        publication_date = INTERVAL_PUBLICATION_DATES.get(extracted_period_interval)
-
-        # Get expected price cap period
-        now = datetime.now()
-        latest_interval = max(
-            (k for k, v in INTERVAL_PUBLICATION_DATES.items() if v <= now),
-            key=INTERVAL_PUBLICATION_DATES.get,
-        )
-        latest_publication_date = INTERVAL_PUBLICATION_DATES[latest_interval]
-
-        valid = publication_date == latest_publication_date
-
-        return ValidationResult(
-            passes=valid,
-            message=f"Expected publication date: {latest_publication_date.strftime(format='%d-%m-%Y')}, "
-            "saw publication date {publication_date.strftime(format='%d-%m-%Y')}",
-        )
 
 
 @check_output_custom(LatestPriceCapValidator(PRICE_CAP_PERIOD_PUBLICATION_DATES))
